Creator Section (almost Done)

+++ app/Http/Controllers/SurveysController.php
@@ -4,9 +4,10 @@ namespace App\Http\Controllers;
 
 use App\Survey;
 use Illuminate\Http\Request;
+use Illuminate\Validation\Rule;
 use App\Http\Requests\StoreSurveyRequest;
-use App\Http\Resources\Survey as SurveyResource;
 use Symfony\Component\HttpFoundation\Response;
+use App\Http\Resources\Survey as SurveyResource;
 
 class SurveysController extends Controller
 {
@@ -29,6 +30,8 @@ class SurveysController extends Controller
      */
     public function store(StoreSurveyRequest $request)
     {
+        $this->authorize('create', Survey::class);
+
         $survey = Survey::create($request->all());
         return (new SurveyResource($survey))->response()->setStatusCode(Response::HTTP_CREATED);
     }
@@ -53,7 +56,25 @@ class SurveysController extends Controller
      */
     public function update(StoreSurveyRequest $request, Survey $survey)
     {
+        $this->authorize('update', $survey);
+
         $survey->update($request->all());
+        return (new SurveyResource($survey))->response()->setStatusCode(Response::HTTP_OK);
+    }
+
+    public function changeStatus(Survey $survey)
+    {
+        $data = request()->validate([
+            'status' => [
+                'required',
+                Rule::in(['draft', 'ready', 'finished'])
+            ]
+        ]);
+
+        $survey->update([
+            'status' => $data['status']
+        ]);
+
         return (new SurveyResource($survey))->response()->setStatusCode(Response::HTTP_OK);
     }
 
@@ -65,6 +86,8 @@ class SurveysController extends Controller
      */
     public function destroy(Survey $survey)
     {
+        $this->authorize('delete', $survey);
+
         $survey->delete();
         return response([], Response::HTTP_NO_CONTENT);
     }
